Batch of styling changes: The Hash choice box, label and button used in file-check-view.fxml had a tendency to stick together, they have been put into an HBox with spacing built in to fix this issue.
The intended styling for the Settings,Minimize and Close buttons were not being properly applied, making them appear too dark and hard to see and/or too light in the other themes, this has been fixed.

The styling for the Selected Tab was not applying properly, making the text blend in, this has been fixed by making it work as intended alongside making the text bold for extra effect.

The anchors which contain the elements for the File Hash Check and Password Generator now are styled properly, making them distinct from the background.

Prepared another anchor in file-check-view.fxml that will house a disclaimer and tutorial on how to use the functionality.

TODO: The accessible-theme.css will need a lot more work to make it work as intended and will require modifications other than just flipping colour values to achieve proper functionality.